Add schema to metadata.json at the end of validation
